requirements.txt
langchain 
langchain-community
chromadb
google-generativeai 
langchain-google-genai 
python-dotenv 
pypdf